add autoload and put user in a class
