+++ App/App.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App;
+
+class App
+{
+    public function run() {
+        session_start();
+
+        require ROOT.'/core/Autoloader.php';
+        $autoloader = new \Core\Autoloader();
+        $autoloader->register();
+    }
+}
