+++ App/Controller/frontend_controller.php
@@ -1,0 +1,17 @@
+<?php
+
+    function main_page() {
+        require('../App/View/frontend/main_page.php');
+    }
+
+    function signup_page() {
+        require('../App/View/frontend/signup.php');
+    }
+
+    function login_page() {
+        require('../App/View/frontend/login.php');
+    }
+
+    function settings_page() {
+        require('../App/View/frontend/settings.php');
+    }
