+++ App/View/frontend/login.php
@@ -1,19 +1,11 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-    <meta charset="utf-8">
-    <title></title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-wEmeIV1mKuiNpC+IOBjI7aAzPcEZeedi5yW5f2yOq55WWLwNGmvvx4Um1vskeMj0" crossorigin="anonymous">
-    <link href="style.css" rel="stylesheet" />
-</head>
-<body>
-    <div class="container">
-        <h1>Se connecter</h1>
-        <?php
-        if (isset($_GET['wrong_credential']) && $_GET['wrong_credential'] == 1) {?>
-            <div class="alert alert-danger" role="alert">
-                Les identifiants ne sont pas corrects!
-            </div>
+<?php
+$page_title = 'Se connecter';
+ob_start();
+
+if (isset($_GET['wrong_credential']) && $_GET['wrong_credential'] == 1) {?>
+    <div class="alert alert-danger" role="alert">
+        Les identifiants ne sont pas corrects!
+    </div>
             <?php
         }
         ?>
@@ -40,8 +32,8 @@
         </form>
         <a href="index.php">Annuler</a>
         <p>Pas encore de compte ? <a href="index.php?action=signup">Créer un compte</a></p>
-    </div>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-p34f1UUtsS3wqzfto5wAAmdvj+osOnFyQFpp4Ua3gs/ZVWx6oOypYoCJhGGScy+8" crossorigin="anonymous"></script>
-</body>
-</html>
+<?php
+$page_body = ob_get_clean();
+require('template.php');
+?>
