+++ App/View/frontend/template.php
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
-<html lang="fr">
+<html lang="en">
 <head>
     <meta charset="utf-8">
-    <title></title>
+    <title>Jonatan Buzek Blog</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-wEmeIV1mKuiNpC+IOBjI7aAzPcEZeedi5yW5f2yOq55WWLwNGmvvx4Um1vskeMj0" crossorigin="anonymous">
     <link href="public/css/style.css" rel="stylesheet" />
 </head>
@@ -29,23 +29,18 @@
                         <a class="nav-link" href="#">CV</a>
                     </li>
                     <?php if (isset($_SESSION['username'])) { ?>
-                    <li class="nav-item">
-                        <a class="nav-link" href="index.php?action=settings">Settings</a>
-                    </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="index.php?action=settings">Settings</a>
+                        </li>
                     <?php } ?>
                 </ul>
             </div>
         </div>
     </nav>
     <div class="container">
-        <h1 class="text-center">Привет мир</h1>
+        <h1 class="text-center"><?php echo $page_title; ?></h1>
         <?php
-        if (isset($_GET['add']) && $_GET['add'] == 1) { ?>
-            <div class="alert alert-success" role="alert">
-                Votre compte a bien été créé
-            </div>
-            <?php
-        }
+        echo $page_body;
         ?>
     </div>
 
@@ -53,18 +48,16 @@
         <div class="text-center p-3" style="background-color: rgba(0, 0, 0, 0.05);">
             © 2021 Copyright: Jonatan Buzek
             <?php
-                if (isset($_SESSION['username']) && !empty($_SESSION['username'])) {
-                    echo '<br>Vous êtes connecté en tant que <span style="color: black">' . ucfirst($_SESSION['username']) . '</span>';
-                    echo '<br><a href="index.php?action=logout" id="logout_link">Logout</a>';
-                } else {
-                    echo '<a href="index.php?action=login" id="login_link">Login</a>';
-                }
+            if (isset($_SESSION['username'])) {
+                echo '<br>Vous êtes connecté en tant que <span style="color: black">' . ucfirst($_SESSION['username']) . '</span>';
+                echo '<br><a href="index.php?action=logout" id="logout_link">Logout</a>';
+            } else {
+                echo '<a href="index.php?action=login" id="login_link">Login</a>';
+            }
             ?>
         </div>
     </footer>
-
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-p34f1UUtsS3wqzfto5wAAmdvj+osOnFyQFpp4Ua3gs/ZVWx6oOypYoCJhGGScy+8" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-p34f1UUtsS3wqzfto5wAAmdvj+osOnFyQFpp4Ua3gs/ZVWx6oOypYoCJhGGScy+8" crossorigin="anonymous"></script>
 </body>
 </html>
