--- controller/frontend_controller.php
@@ -1,17 +1,0 @@
-<?php
-
-function main_page() {
-    require('view/frontend/main_page.php');
-}
-
-function signup_page() {
-    require('view/frontend/signup.php');
-}
-
-function login_page() {
-    require('view/frontend/login.php');
-}
-
-function settings_page() {
-    require('view/frontend/settings.php');
-}
